+++ src/main/java/com/depromeet/archive/infra/user/oauth/provider/KakaoClient.java
@@ -1,11 +1,11 @@
-package com.depromeet.archive.infra.user.provider;
+package com.depromeet.archive.infra.user.oauth.provider;
 
 import com.depromeet.archive.api.dto.user.OAuthRegisterDto;
 import com.depromeet.archive.domain.user.command.OAuthRegisterCommand;
 import com.depromeet.archive.domain.user.entity.OAuthProvider;
 import com.depromeet.archive.exception.user.OAuthRegisterFailException;
-import com.depromeet.archive.infra.user.provider.dto.KakaoProviderRequirements;
-import com.depromeet.archive.infra.user.provider.dto.KakaoUserInfo;
+import com.depromeet.archive.infra.user.oauth.provider.dto.KakaoProviderRequirements;
+import com.depromeet.archive.infra.user.oauth.provider.dto.KakaoUserInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
